add a bunch or reviews and auto reviews to seed file

+++ frontend/src/components/AccountPage/AccountPage.js
@@ -51,6 +51,7 @@ function AccountPage () {
     }
 
     const handleDelete = () => {
+        if (sessionUser.email === "demo@demo.com") {return alert("Demo User cannot be deleted")}
         if (window.confirm("Are you sure you want to delete your account?")) {
             dispatch(deleteUser(sessionUser.id));
         }
